De-duplicate common note actions
