+++ Assets/__Scripts/Editor/Input/BeatmapSharedNoteInputController.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: ace5f9ab0d5c4840bd635138bec0f7f3
+timeCreated: 1781084402
